--- src/storage/loaders/db.ts
@@ -1,5 +1,0 @@
-import BookmarksService from '../services/bookmarks';
-
-export default async () => {
-  await BookmarksService.load();
-};
