--- src/storage/loaders/index.ts
@@ -1,7 +1,0 @@
-import db from './db';
-import communication from './communication';
-
-export default async () => {
-  await db();
-  communication();
-};
