+++ src/storage/services/bookmarks.ts
@@ -55,7 +55,7 @@ class BookmarksService extends BookmarksServiceBase {
     return Math.max(0, Math.min(index ?? children.length, children.length));
   }
 
-  public async load() {
+  public async start() {
     const { roots } = await readJsonFile<IBookmarksDocument>(
       config.bookmarks,
       config.default.bookmarks,
