+++ src/storage/services/db.ts
@@ -1,0 +1,35 @@
+import * as sqlite from 'better-sqlite3';
+import { Database } from 'better-sqlite3';
+import { promises as fs } from 'fs';
+
+import { pathExists } from '~/common/utils/files';
+import { config } from '../constants';
+import BookmarksService from './bookmarks';
+
+class DbService {
+  public history: Database;
+
+  public async start() {
+    this.history = await this.create(config.history, config.default.history);
+
+    await BookmarksService.start();
+  }
+
+  private async create(path: string, schemaPath: string) {
+    const exists = await pathExists(path);
+
+    const db = sqlite(path, { verbose: console.log });
+
+    if (!exists) {
+      const schema = await fs.readFile(schemaPath, 'utf8');
+
+      db.exec(schema);
+
+      console.log(`History database created at ${path}`);
+    }
+
+    return db;
+  }
+}
+
+export default new DbService();
